+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # 14.3.0
 
  * Genesis and Express firmware updates.
+ * Ability to use more than one variable per sequence.
+ * Ability for Lua developers to see current point group via `__GROUP__` sequence variable.
 
 # 14.2.2